refactor: move and rename enumerations
